MAINT: Move type predicates into dedicated module.
Break out `iscorescalar`, `iscoresequence`, and `iscoretype` into a module
dedicated to type checking.

Putting these functions in their own module helps avoid an import cycle while
breaking out other components from the interactive module.
