+++ blaze/tests/test_interactive.py
@@ -22,9 +22,6 @@ from blaze.interactive import (
     concrete_head,
     data,
     expr_repr,
-    iscorescalar,
-    iscoresequence,
-    iscoretype,
     to_html,
 )
 from blaze.utils import tmpfile, example, normalize
@@ -564,40 +561,3 @@ def test_isidentical_regr():
                            np.ndarray)])
 def test_coerce_core(data, dshape, exp_type):
     assert isinstance(coerce_core(data, dshape), exp_type)
-
-
-@pytest.mark.parametrize('data,res',
-                         [(1, True),
-                          (1.1, True),
-                          ("foo", True),
-                          ([1, 2], False),
-                          ((1, 2), False),
-                          (pd.Series([1, 2]), False)])
-def test_iscorescalar(data, res):
-    assert iscorescalar(data) == res
-
-
-@pytest.mark.parametrize('data,res',
-                         [(1, False),
-                          ("foo", False),
-                          ([1, 2], True),
-                          ((1, 2), True),
-                          (pd.Series([1, 2]), True),
-                          (pd.DataFrame([[1, 2], [3, 4]]), True),
-                          (np.ndarray([1, 2]), True),
-                          (into(da.core.Array, [1, 2], chunks=(10,)), False)])
-def test_iscoresequence(data, res):
-    assert iscoresequence(data) == res
-
-
-@pytest.mark.parametrize('data,res',
-                         [(1, True),
-                          ("foo", True),
-                          ([1, 2], True),
-                          ((1, 2), True),
-                          (pd.Series([1, 2]), True),
-                          (pd.DataFrame([[1, 2], [3, 4]]), True),
-                          (np.ndarray([1, 2]), True),
-                          (into(da.core.Array, [1, 2], chunks=(10,)), False)])
-def test_iscoretype(data, res):
-    assert iscoretype(data) == res
